+++ package.json
@@ -52,6 +52,7 @@
     "rehype-slug": "^6.0.0",
     "resend": "^3.2.0",
     "schema-dts": "^1.1.2",
+    "server-only": "^0.0.1",
     "shiki": "^1.3.0",
     "valibot": "^0.30.0",
     "zod": "^3.22.4"
