+++ src/lib/actions.tsx
@@ -1,6 +1,7 @@
 'use server';
 
 import { unstable_noStore as noStore } from 'next/cache';
+import { ErrorResponse } from 'resend';
 import { parse } from 'valibot';
 
 import { ContactFormAcknowledgement } from '@/components/emails/contact-form-acknowledgement';
@@ -9,8 +10,7 @@ import { email, firstName, lastName } from '@/constants/content';
 import { env } from '@/env.mjs';
 import { prisma, resend } from '@/lib/clients';
 
-import { CreateEmailError } from './utils';
-import { ContactForm, contactForm } from './validation-schema';
+import { ContactForm, contactForm } from './validation';
 
 export const incrBlogViewById = async (id: string) => {
   noStore();
@@ -56,3 +56,12 @@ export const submitContactForm = async (data: ContactForm) => {
     );
   }
 };
+
+class CreateEmailError extends Error {
+  error: ErrorResponse;
+
+  constructor(message: string, error: ErrorResponse) {
+    super(message);
+    this.error = error;
+  }
+}
