+++ src/lib/json-ld.ts
@@ -1,4 +1,3 @@
-import { ErrorResponse } from 'resend';
 import { Person } from 'schema-dts';
 
 import {
@@ -17,7 +16,7 @@ import {
   getPlatformProfiles,
 } from './queries';
 
-export const getJsonLdPerson = async () => {
+export const getPerson = async () => {
   const [latestJobTitle, personalPhoto, platformProfiles] = await Promise.all([
     getExperiences().then((experience) => experience[0]?.jobTitle),
     getPersonalPhoto(),
@@ -39,15 +38,3 @@ export const getJsonLdPerson = async () => {
     description: selfIntroduction,
   } satisfies Person;
 };
-
-export const encodeHtmlEntities = (str: string) =>
-  str.replace(/[\u00A0-\u9999<>&]/gim, (i) => `&#${i.charCodeAt(0)};`);
-
-export class CreateEmailError extends Error {
-  error: ErrorResponse;
-
-  constructor(message: string, error: ErrorResponse) {
-    super(message);
-    this.error = error;
-  }
-}
